fix: MainPage를 중첩으로 개발
